Create internal flavor
