+++ app/build.gradle
@@ -75,6 +75,9 @@ android {
     }
     flavorDimensions "store"
     productFlavors {
+        internal{
+            dimension "store"
+        }
         fdroid{
             dimension "store"
         }
@@ -84,8 +87,8 @@ android {
     }
     variantFilter { variant ->
         def names = variant.flavors*.name
-        if (names.contains("fdroid") && variant.buildType.name == "debug") {
-            // We only need the F-Droid release version
+        if ((names.contains("fdroid") || names.contains("internal")) && variant.buildType.name == "debug") {
+            // We only need release version for F-Droid and internal builds
             setIgnore(true)
         }
     }
