+++ app/src/main/java/com/duckduckgo/app/referral/AppInstallationReferrerParser.kt
@@ -17,7 +17,10 @@
 package com.duckduckgo.app.referral
 
 import com.duckduckgo.app.referral.ParsedReferrerResult.*
+import com.duckduckgo.di.scopes.AppObjectGraph
+import com.squareup.anvil.annotations.ContributesBinding
 import timber.log.Timber
+import javax.inject.Inject
 
 interface AppInstallationReferrerParser {
 
@@ -25,7 +28,8 @@ interface AppInstallationReferrerParser {
 }
 
 @Suppress("SameParameterValue")
-class QueryParamReferrerParser : AppInstallationReferrerParser {
+@ContributesBinding(AppObjectGraph::class)
+class QueryParamReferrerParser @Inject constructor() : AppInstallationReferrerParser {
 
     override fun parse(referrer: String): ParsedReferrerResult {
         val referrerParts = splitIntoConstituentParts(referrer)
