+++ app/src/main/java/com/duckduckgo/app/referral/AppInstallationReferrerStateListener.kt
@@ -16,6 +16,11 @@
 
 package com.duckduckgo.app.referral
 
+import com.duckduckgo.app.statistics.AtbInitializerListener
+import timber.log.Timber
+import javax.inject.Inject
+import javax.inject.Singleton
+
 interface AppInstallationReferrerStateListener {
 
     fun initialiseReferralRetrieval()
@@ -25,4 +30,39 @@ interface AppInstallationReferrerStateListener {
         const val MAX_REFERRER_WAIT_TIME_MS = 1_500L
     }
 
+}
+
+@Singleton
+class EmptyReferrerStateListener @Inject constructor() : AppInstallationReferrerStateListener, AtbInitializerListener {
+
+    private var referralResult: ParsedReferrerResult = ParsedReferrerResult.ReferrerInitialising
+
+    /**
+     * Initialises the referrer service. This should only be called once.
+     */
+    override fun initialiseReferralRetrieval() {
+        Timber.d("Empty referrer, nothing to do here")
+        referralResult = ParsedReferrerResult.ReferrerNotFound()
+    }
+
+    /**
+     * Retrieves the app installation referral code.
+     * This might return a result immediately or might wait for a result to become available. There is no guarantee that a result will ever be returned.
+     *
+     * It is the caller's responsibility to guard against this function not returning a result in a timely manner, or not returning a result ever.
+     */
+    override suspend fun waitForReferrerCode(): ParsedReferrerResult {
+        if (referralResult != ParsedReferrerResult.ReferrerInitialising) {
+            Timber.d("Referrer already determined; immediately answering")
+            return referralResult
+        }
+
+        return referralResult
+    }
+
+    override suspend fun beforeAtbInit() {
+        waitForReferrerCode()
+    }
+
+    override fun beforeAtbInitTimeoutMillis(): Long = AppInstallationReferrerStateListener.MAX_REFERRER_WAIT_TIME_MS
 }
