+++ app/src/play/java/com/duckduckgo/app/di/PlayHttpsPersisterModule.kt
@@ -21,13 +21,19 @@ import com.duckduckgo.app.global.store.BinaryDataStore
 import com.duckduckgo.app.httpsupgrade.store.HttpsBloomFilterSpecDao
 import com.duckduckgo.app.httpsupgrade.store.HttpsDataPersister
 import com.duckduckgo.app.httpsupgrade.store.HttpsEmbeddedDataPersister
+import com.duckduckgo.di.scopes.AppObjectGraph
 import com.duckduckgo.httpsupgrade.store.PlayHttpsEmbeddedDataPersister
+import com.squareup.anvil.annotations.ContributesTo
 import com.squareup.moshi.Moshi
 import dagger.Module
 import dagger.Provides
 
 @Module
-class HttpsPersisterModule {
+@ContributesTo(
+    scope = AppObjectGraph::class,
+    replaces = [HttpsPersisterModule::class]
+)
+class PlayHttpsPersisterModule {
 
     @Provides
     fun providesPlayHttpsEmbeddedDataPersister(
